Модуляция (#5)
* 1

* Mодуляция

* Поправила модульность

* удалена папка модуль, файлы в корне js
